Move MarketStatsCheck to a separate component
